feat: login submission hook have been implemented
