+++ frontend/src/pages/Login/index.tsx
@@ -1,29 +1,17 @@
-import React, { useState } from "react";
-import { useNavigate } from "react-router-dom";
-import axios from "axios";
+import React from "react";
 import DescriptionPage from "../../components/DescriptionTemplate";
+import useLogin from "../../hooks/useLogin";
 import './styles.css';
 
 const Login: React.FC = () => {
-    const [email, setEmail] = useState("");
-    const [password, setPassword] = useState("");
-    const [errorMessage, setErrorMessage] = useState("");
-
-    const navigate = useNavigate();
-
-    const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-        event.preventDefault();
-        try {
-            const response = await axios.post("http://localhost:3333/users/login", {
-                email: email,
-                password: password
-            });
-            document.cookie = `jwt=${response.data.token}; path=/;`;
-            navigate("/"); // Redirecionar para a página inicial após o login        
-        } catch (error) {
-            setErrorMessage("Email ou senha inválidos. Por favor, tente novamente.");
-        }
-    };
+    const {
+        email,
+        setEmail,
+        password,
+        setPassword,
+        errorMessage,
+        handleSubmit
+    } = useLogin();
 
     return (
         <div className="login-page">
